+++ src/main/java/com/solegendary/reignofnether/unit/units/monsters/CreeperUnit.java
@@ -120,7 +120,7 @@ public class CreeperUnit extends Creeper implements Unit, AttackerUnit {
 
     // combat stats
     public boolean getWillRetaliate() {return willRetaliate;}
-    public int getAttackCooldown() {return (int) ((20 / attacksPerSecond) * getAttackSlowdownMultiplier());}
+    public float getAttackCooldown() {return ((20 / attacksPerSecond) * getAttackSlowdownMultiplier());}
     public float getAttacksPerSecond() {return 20f / getAttackCooldown();}
     public float getBaseAttacksPerSecond() {return attacksPerSecond;}
     public float getAggroRange() {return aggroRange;}
@@ -239,7 +239,7 @@ public class CreeperUnit extends Creeper implements Unit, AttackerUnit {
         this.moveGoal = new MoveToTargetBlockGoal(this, false, 0);
         this.targetGoal = new SelectedTargetGoal<>(this, true, true);
         this.garrisonGoal = new GarrisonGoal(this);
-        this.attackGoal = new CreeperAttackUnitGoal(this, getAttackCooldown(), false);
+        this.attackGoal = new CreeperAttackUnitGoal(this, (int) getAttackCooldown(), false);
     }
 
     @Override
