+++ src/main/java/com/solegendary/reignofnether/unit/units/monsters/ZombieUnit.java
@@ -116,7 +116,7 @@ public class ZombieUnit extends Zombie implements Unit, AttackerUnit, Convertabl
 
     // combat stats
     public boolean getWillRetaliate() {return willRetaliate;}
-    public int getAttackCooldown() {return (int) ((20 / attacksPerSecond) * getAttackSlowdownMultiplier());}
+    public float getAttackCooldown() {return ((20 / attacksPerSecond) * getAttackSlowdownMultiplier());}
     public float getAttacksPerSecond() {return 20f / getAttackCooldown();}
     public float getBaseAttacksPerSecond() {return attacksPerSecond;}
     public float getAggroRange() {return aggroRange;}
