+++ src/main/java/com/solegendary/reignofnether/unit/units/neutral/PandaUnit.java
@@ -173,7 +173,7 @@ public class PandaUnit extends Panda implements Unit, AttackerUnit {
         return willRetaliate;
     }
 
-    public int getAttackCooldown() {return (int) ((20 / attacksPerSecond) * getAttackSlowdownMultiplier());}
+    public float getAttackCooldown() {return ((20 / attacksPerSecond) * getAttackSlowdownMultiplier());}
     public float getAttacksPerSecond() {return 20f / getAttackCooldown();}
     public float getBaseAttacksPerSecond() {return attacksPerSecond;}
