+++ src/main/java/com/solegendary/reignofnether/unit/units/neutral/WolfUnit.java
@@ -114,7 +114,7 @@ public class WolfUnit extends Wolf implements Unit, AttackerUnit {
     @Nullable
     public ResourceCost getCost() {return ResourceCosts.WOLF;}
     public boolean getWillRetaliate() {return willRetaliate;}
-    public int getAttackCooldown() {return (int) ((20 / attacksPerSecond) * getAttackSlowdownMultiplier());}
+    public float getAttackCooldown() {return ((20 / attacksPerSecond) * getAttackSlowdownMultiplier());}
     public float getAttacksPerSecond() {return 20f / getAttackCooldown();}
     public float getBaseAttacksPerSecond() {return attacksPerSecond;}
     public float getAggroRange() {return aggroRange;}
