+++ src/main/java/com/solegendary/reignofnether/unit/units/piglins/GruntUnit.java
@@ -134,7 +134,7 @@ public class GruntUnit extends Piglin implements Unit, WorkerUnit, AttackerUnit,
     @Nullable
     public ResourceCost getCost() {return ResourceCosts.GRUNT;}
     public boolean getWillRetaliate() {return willRetaliate;}
-    public int getAttackCooldown() {return (int) ((20 / attacksPerSecond) * getAttackSlowdownMultiplier());}
+    public float getAttackCooldown() {return ((20 / attacksPerSecond) * getAttackSlowdownMultiplier());}
     public float getAttacksPerSecond() {return 20f / getAttackCooldown();}
     public float getBaseAttacksPerSecond() {return attacksPerSecond;}
     public float getAggroRange() {return aggroRange;}
